add unknown inputs with zod validation on backend

+++ src/actions/actions.ts
@@ -6,9 +6,9 @@ import { Pet } from '@prisma/client';
 import prisma from '@/lib/db';
 import { sleep } from '@/lib/utils';
 import { PetEssentials } from '@/lib/types';
-import { petFormSchema } from '@/lib/validations';
+import { petFormSchema, petIdSchema } from '@/lib/validations';
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
 
   const validatedPet = petFormSchema.safeParse(pet);
@@ -31,15 +31,24 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+
+  if (!validatedPetId.success || !validatedPet.success) {
+    return {
+      message: "Invalid pet data.",
+    }
+  }
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -50,13 +59,21 @@ export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function checkoutPet(petId: Pet['id']) {
+export async function checkoutPet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data.",
+    }
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
